% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alley}
\alias{alley}
\title{Undirected edge pairing of alley connecting nodes}
\format{A data frame with 452 rows and 2 variables
\describe{
  \item{x}{The smaller integer of the pair of vertices}
  \item{y}{The larger integer of the pair of vertices}
}}
\usage{
alley
}
\description{
Data used to establish possible connections used by Jack between nodes via alleyways
}
\keyword{datasets}
