% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roads}
\alias{roads}
\title{Undirected edge pairing of roads connecting nodes}
\format{A data frame with 767 rows and 2 variables
\describe{
  \item{x}{The smaller integer of the pair of vertices}
  \item{y}{The larger integer of the pair of vertices}
}}
\usage{
roads
}
\description{
Data used to establish possible connections used by Jack between nodes
}
\keyword{datasets}
