% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_round.R
\name{start_round}
\alias{start_round}
\title{Start a new round}
\usage{
start_round(initial_murder)
}
\arguments{
\item{initial_murder}{integer Space of the initial murder(s)}
}
\description{
Generate the initial list for a new round
}
\details{
export
}
