% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_a_step.R
\name{take_a_carriage}
\alias{take_a_carriage}
\title{Track carriage movement}
\usage{
take_a_carriage(paths)
}
\arguments{
\item{paths}{list of all possible paths already traveled}
}
\value{
list of all possible paths traveled by Jack
}
\description{
Track two steps of unknown movement by Jack, on roads
}
\examples{
possibilities = start_round(64)
possibilities = take_a_carriage(possibilities)
}
