% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_a_step.R
\name{trim_possibilities}
\alias{trim_possibilities}
\title{Trim possible paths}
\usage{
trim_possibilities(paths, node)
}
\arguments{
\item{paths}{list of all possible paths already traveled}

\item{node}{vector of length 1 or 2 which specifies blocked nodes due to the presence of a policeman}
}
\value{
list of trimmed possible paths traveled by Jack
}
\description{
Remove known impossible end points for Jack, typically as a result of having found, but not arrested Jack.
}
\examples{
possibilities = start_round(64)
possibilities = take_a_carriage(possibilities)
possibilities = trim_possibilities(possibilities,82)
possibilities = trim_possibilities(possibilities,c(66,67))
}
