% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_space.R
\name{inspect_space}
\alias{inspect_space}
\title{Update paths based on inspections}
\usage{
inspect_space(paths, space, clue)
}
\arguments{
\item{paths}{list of all possible paths already traveled}

\item{space}{vector of integers of the spaces inspected}

\item{clue}{single logical value indicating if evidence of Jack was found}
}
\value{
list of all possible paths traveled by Jack
}
\description{
Updated the list of possible paths based on the results of police investigation
}
